# Non-recursive fibonacci implementation.