--- a/salt/modules/test.py
+++ b/salt/modules/test.py
@@ -326,16 +326,15 @@ def fib(num):
     start = time.time()
     if num < 2:
         return num, time.time() - start
-    return _fib(num-1) + _fib(num-2), time.time() - start
 
+    prev = 0
+    curr = 1
+    i = 1
+    while i < num:
+        prev, curr = curr, prev + curr
+        i += 1
 
-def _fib(num):
-    '''
-    Helper method for test.fib, doesn't calculate the time.
-    '''
-    if num < 2:
-        return num
-    return _fib(num-1) + _fib(num-2)
+    return curr, time.time() - start
 
 
 def collatz(start):