### What does this PR do?
Non-recursive `test.fib` implementation.
- faster
- unlimited

### What issues does this PR fix or reference?
#41042

### Previous Behavior
```
$ python -m timeit -s 'import salt.modules.test as t' 't.fib_old(10)'
100000 loops, best of 3: 18.4 usec per loop
$ python -m timeit -s 'import salt.modules.test as t' 't.fib_old(20)'
100 loops, best of 3: 2.21 msec per loop
```

### New Behavior
```
$ python -m timeit -s 'import salt.modules.test as t' 't.fib(10)'
1000000 loops, best of 3: 1.07 usec per loop
$ python -m timeit -s 'import salt.modules.test as t' 't.fib(20)'
1000000 loops, best of 3: 1.58 usec per loop
```

### Tests written?
No