@thatch45 2017-05-05
At least there is no complaining about my implementation of the collatz conjecture :)