@basepi 2017-05-08
For the record, I purposely broke this years ago to make it slower, because it wasn't actually useful for showing delays in jobs. Turns out fibonacci is not hard to calculate to very high values in milliseconds.