@thatch45 2017-05-08
I am honestly amazed that this has seen so much attention, no love for the collatz conjecture